requirements-dev.txt
-r requirements.txt
httpx>=0.28
playwright>=1.48
beautifulsoup4>=4.12
google-generativeai>=0.8
pytest>=8.0
pytest-asyncio>=0.24
ruff>=0.8
